Move Dice and RollResult to their own files
