+++ client/Commands/CombatantCommander.ts
@@ -5,7 +5,8 @@ import { probablyUniqueString } from "../../common/Toolbox";
 import { Combatant } from "../Combatant/Combatant";
 import { CombatantDetails } from "../Combatant/CombatantDetails";
 import { CombatantViewModel } from "../Combatant/CombatantViewModel";
-import { Dice, RollResult } from "../Rules/Rules";
+import { Dice } from "../Rules/Dice";
+import { RollResult } from "../Rules/RollResult";
 import { CurrentSettings } from "../Settings/Settings";
 import { TrackerViewModel } from "../TrackerViewModel";
 import { Metrics } from "../Utility/Metrics";
