+++ client/TextEnricher/TextEnricher.tsx
@@ -11,7 +11,8 @@ import {
 import { Listing } from "../Library/Listing";
 import { SpellLibrary } from "../Library/SpellLibrary";
 import { Conditions } from "../Rules/Conditions";
-import { Dice, IRules } from "../Rules/Rules";
+import { Dice } from "../Rules/Dice";
+import { IRules } from "../Rules/Rules";
 
 interface ReplaceConfig {
   [name: string]: {
